render widget after mount & update deps
