+++ .changeset/curly-socks-worry.md
@@ -1,0 +1,5 @@
+---
+"@superfluid-finance/widget": patch
+---
+
+Render widget after mount (no SSR)
