+++ .changeset/mean-hounds-yawn.md
@@ -1,0 +1,5 @@
+---
+"@superfluid-finance/widget": patch
+---
+
+Update deps
