+++ apps/hosted-widget/package.json
@@ -15,10 +15,10 @@
     "@emotion/react": "^11.11.1",
     "@emotion/server": "^11.11.0",
     "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.14.12",
-    "@mui/material": "^5.14.12",
-    "@segment/analytics-next": "^1.56.0",
-    "@sentry/nextjs": "^7.72.0",
+    "@mui/icons-material": "^5.14.15",
+    "@mui/material": "^5.14.15",
+    "@segment/analytics-next": "^1.59.0",
+    "@sentry/nextjs": "^7.75.1",
     "@superfluid-finance/widget": "workspace:*",
     "@web3modal/ethereum": "^2.7.1",
     "@web3modal/react": "^2.7.1",
@@ -27,16 +27,16 @@
     "react": "^18.2.0",
     "react-blockies": "^1.4.1",
     "react-dom": "^18.2.0",
-    "viem": "^1.16.2",
-    "wagmi": "^1.4.3",
+    "viem": "^1.17.1",
+    "wagmi": "^1.4.5",
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/node": "^20.8.4",
-    "@types/react": "^18.2.27",
-    "@types/react-blockies": "^1.4.2",
-    "eslint": "^8.51.0",
-    "eslint-config-next": "^13.5.4",
+    "@types/node": "^20.8.9",
+    "@types/react": "^18.2.33",
+    "@types/react-blockies": "^1.4.3",
+    "eslint": "^8.52.0",
+    "eslint-config-next": "^13.5.6",
     "typescript": "^5.2.2"
   }
 }
