+++ apps/widget-builder/package.json
@@ -15,14 +15,14 @@
     "@emotion/react": "^11.11.1",
     "@emotion/server": "^11.11.0",
     "@emotion/styled": "^11.11.0",
-    "@faker-js/faker": "^8.1.0",
+    "@faker-js/faker": "^8.2.0",
     "@monaco-editor/react": "^4.6.0",
-    "@mui/icons-material": "^5.14.12",
+    "@mui/icons-material": "^5.14.15",
     "@mui/lab": "5.0.0-alpha.147",
-    "@mui/material": "^5.14.12",
+    "@mui/material": "^5.14.15",
     "@pinata/sdk": "^2.1.0",
-    "@segment/analytics-next": "^1.56.0",
-    "@sentry/nextjs": "^7.72.0",
+    "@segment/analytics-next": "^1.59.0",
+    "@sentry/nextjs": "^7.75.1",
     "@superfluid-finance/widget": "workspace:*",
     "@web3modal/ethereum": "^2.7.1",
     "@web3modal/react": "^2.7.1",
@@ -30,28 +30,28 @@
     "lodash": "^4.17.21",
     "lru-cache": "^10.0.1",
     "monaco-editor": "^0.44.0",
-    "mui-color-input": "^2.0.0",
+    "mui-color-input": "^2.0.1",
     "next": "13.4.19",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-google-recaptcha": "^3.1.0",
     "react-hook-form": "~7.47.0",
     "request-ip": "^3.3.0",
-    "viem": "^1.16.2",
-    "wagmi": "^1.4.3",
+    "viem": "^1.17.1",
+    "wagmi": "^1.4.5",
     "zod": "^3.22.4",
     "zod-to-json-schema": "^3.21.4"
   },
   "devDependencies": {
     "@hookform/devtools": "^4.3.1",
-    "@types/lodash": "^4.14.199",
-    "@types/node": "^20.8.4",
-    "@types/react": "^18.2.27",
-    "@types/react-dom": "^18.2.12",
-    "@types/react-google-recaptcha": "^2.1.6",
+    "@types/lodash": "^4.14.200",
+    "@types/node": "^20.8.9",
+    "@types/react": "^18.2.33",
+    "@types/react-dom": "^18.2.14",
+    "@types/react-google-recaptcha": "^2.1.7",
     "@types/request-ip": "^0.0.39",
-    "eslint": "^8.51.0",
-    "eslint-config-next": "^13.5.4",
+    "eslint": "^8.52.0",
+    "eslint-config-next": "^13.5.6",
     "typescript": "^5.2.2"
   }
 }
