+++ examples/b2b-service-demo/package.json
@@ -10,7 +10,7 @@
     "tsc": "tsc --noEmit"
   },
   "dependencies": {
-    "@sentry/nextjs": "^7.72.0",
+    "@sentry/nextjs": "^7.75.1",
     "@superfluid-finance/widget": "workspace:*",
     "@web3modal/ethereum": "^2.7.1",
     "@web3modal/react": "^2.7.1",
@@ -18,8 +18,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-use-intercom": "^5.1.4",
-    "viem": "^1.16.2",
-    "wagmi": "^1.4.3",
+    "viem": "^1.17.1",
+    "wagmi": "^1.4.5",
     "zod": "^3.22.4"
   },
   "devDependencies": {
