+++ examples/widget-vite-react-rainbowkit/package.json
@@ -11,24 +11,24 @@
     "tsc": "tsc --noEmit"
   },
   "dependencies": {
-    "@rainbow-me/rainbowkit": "^1.1.1",
+    "@rainbow-me/rainbowkit": "^1.1.3",
     "@superfluid-finance/widget": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "viem": "^1.16.2",
-    "wagmi": "^1.4.3",
+    "viem": "^1.17.1",
+    "wagmi": "^1.4.5",
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/react": "^18.2.27",
-    "@types/react-dom": "^18.2.12",
-    "@typescript-eslint/eslint-plugin": "^6.7.5",
-    "@typescript-eslint/parser": "^6.7.5",
+    "@types/react": "^18.2.33",
+    "@types/react-dom": "^18.2.14",
+    "@typescript-eslint/eslint-plugin": "^6.9.0",
+    "@typescript-eslint/parser": "^6.9.0",
     "@vitejs/plugin-react": "^4.1.0",
-    "eslint": "^8.51.0",
+    "eslint": "^8.52.0",
     "eslint-plugin-react-hooks": "5.0.0-canary-41f0e9dae-20230907",
     "eslint-plugin-react-refresh": "^0.4.3",
     "typescript": "^5.2.2",
-    "vite": "^4.4.11"
+    "vite": "^4.5.0"
   }
 }
