+++ package.json
@@ -28,7 +28,7 @@
   },
   "devDependencies": {
     "@changesets/cli": "^2.26.2",
-    "eslint": "^8.51.0",
+    "eslint": "^8.52.0",
     "eslint-config-custom": "workspace:*",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-plugin-unused-imports": "^3.0.0",
@@ -36,7 +36,7 @@
     "lint-staged": "^14.0.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.0.3",
-    "turbo": "^1.10.15",
+    "turbo": "^1.10.16",
     "typescript": "^5.2.2"
   },
   "packageManager": "pnpm@8.6.1",
