+++ packages/eslint-config-custom/package.json
@@ -5,9 +5,9 @@
   "license": "MIT",
   "main": "index.js",
   "dependencies": {
-    "eslint-config-next": "^13.5.4",
+    "eslint-config-next": "^13.5.6",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-config-turbo": "^1.10.15",
+    "eslint-config-turbo": "^1.10.16",
     "eslint-plugin-react": "^7.33.2"
   },
   "publishConfig": {
