+++ packages/widget/package.json
@@ -61,19 +61,19 @@
   "dependencies": {
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@hookform/resolvers": "^3.3.1",
-    "@mui/icons-material": "^5.14.12",
+    "@hookform/resolvers": "^3.3.2",
+    "@mui/icons-material": "^5.14.15",
     "@mui/lab": "5.0.0-alpha.147",
-    "@mui/utils": "^5.14.12",
-    "@superfluid-finance/metadata": "^1.1.17",
-    "@superfluid-finance/tokenlist": "^3.2.0",
+    "@mui/utils": "^5.14.15",
+    "@superfluid-finance/metadata": "^1.1.18",
+    "@superfluid-finance/tokenlist": "^3.3.2",
     "abitype": "^0.9.10",
     "blockies-ts": "^1.0.0",
     "immer": "^10.0.3",
     "lodash.isequal": "^4.5.0",
     "lodash.memoize": "^4.1.2",
     "lodash.uniqwith": "^4.5.0",
-    "nanoid": "^5.0.1",
+    "nanoid": "^5.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-error-boundary": "^4.0.11",
@@ -83,23 +83,23 @@
     "zod-validation-error": "^1.4.0"
   },
   "devDependencies": {
-    "@mui/material": "^5.14.12",
+    "@mui/material": "^5.14.15",
     "@superfluid-finance/ethereum-contracts": "^1.8.1",
     "@testing-library/react": "^14.0.0",
-    "@types/lodash.isequal": "^4.5.6",
-    "@types/lodash.memoize": "^4.1.7",
-    "@types/lodash.uniqwith": "^4.5.7",
-    "@types/react": "^18.2.27",
-    "@types/react-dom": "^18.2.12",
+    "@types/lodash.isequal": "^4.5.7",
+    "@types/lodash.memoize": "^4.1.8",
+    "@types/lodash.uniqwith": "^4.5.8",
+    "@types/react": "^18.2.33",
+    "@types/react-dom": "^18.2.14",
     "@vitejs/plugin-react": "^4.1.0",
     "@wagmi/cli": "^1.5.2",
     "jsdom": "^22.1.0",
     "tsconfig": "workspace:*",
     "typedoc": "^0.25.2",
     "typescript": "^5.2.2",
-    "viem": "^1.16.2",
+    "viem": "^1.17.1",
     "vitest": "^0.34.6",
-    "wagmi": "^1.4.3"
+    "wagmi": "^1.4.5"
   },
   "peerDependencies": {
     "viem": ">=1",
