+++ tests/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "devDependencies": {
     "@currents/playwright": "^0.8.1",
-    "@playwright/test": "^1.38.1",
+    "@playwright/test": "^1.39.0",
     "@synthetixio/synpress": "^3.7.1",
     "dotenv": "16.3.1",
     "ethers": "6.7.1",
